+++ pyvo/mivot/features/sky_coord_builder.py
@@ -6,6 +6,7 @@ Utility transforming MIVOT annotation into SKYCoord instances
 from astropy.coordinates import SkyCoord
 from astropy import units as u
 from astropy.coordinates import ICRS, Galactic, FK4, FK5
+from pyvo.mivot.utils.exceptions import NoMatchingDMType
 
 
 class MangoRoles:
@@ -65,10 +66,16 @@ class SkyCoordBuilder(object):
 
         returns
         -------
-        - SkyCoord instance or None
+        - SkyCoord instance
+
+        raises
+        ------
+        - NoMatchingDMType: if the SkyCoord instance cannot be built.
         """
         if self._mivot_instance_dict and self._mivot_instance_dict["dmtype"] == "mango:EpochPosition":
             return self._build_sky_coord_from_mango()
+        raise NoMatchingDMType("No INSTANCE with dmtype='mango:EpochPosition' has been found:"
+                               " cannot build a SkyCoord from annotations")
 
     def _set_year_time_format(self, hk_field, besselian=False):
         """